Added the HttpResponse class.
The HttpResponse class is a new abstract class that is used by
Responders to send data to the client. The HttpResponseImpl class is an
implementation of it. Responders now receive pointers to
HttpRequest and HttpResponse objects (instead of a pointer to an
HttpConnection) when they may respond to a request.